AGT code
This code for AGT worked to the point that when unconnected to the ESP, it went to the pickup case and retreived GNSS position succesfully and transmitted an iridium message succesfully. GNSS was of by about 200m, so maybe a longer start up and wait time for GNSS would triangulate it better.